requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0

# Vision and image processing
opencv-python>=4.8.0
Pillow>=10.0.0
timm>=0.9.0
albumentations>=1.3.0

# Text processing and NLP
tokenizers>=0.13.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Hugging Face ecosystem
huggingface-hub>=0.15.0
evaluate>=0.4.0

# MoE and efficiency
deepspeed>=0.9.0
fairscale>=0.4.13

# Data science and utilities
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and logging
pyyaml>=6.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Chart-specific processing
camelot-py[cv]>=0.10.1  # Table extraction from PDFs
pdfplumber>=0.9.0       # PDF text extraction
pytesseract>=0.3.10     # OCR
easyocr>=1.7.0          # Advanced OCR

# Evaluation metrics
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.0

# Optional: For advanced features
# flash-attn>=2.0.0     # Uncomment if using FlashAttention
# apex                  # Uncomment if using NVIDIA Apex 